Move functions alongside protocol conformance declaration
